.ipynb_checkpoints/requirements-checkpoint.txt
matplotlib
pillow
scikit-image
numpy
pandas
keras
tensorflow==2.3.1
scikit-learn